+++ design/actdia.go
@@ -1,11 +1,13 @@
 package design
 
-import "github.com/gregoryv/draw/shape"
+import (
+	"github.com/gregoryv/draw/shape"
+)
 
 func NewActivityDiagram() *ActivityDiagram {
 	return &ActivityDiagram{
 		Diagram: NewDiagram(),
-		Spacing: 40,
+		Spacing: 60,
 	}
 }
 
@@ -26,18 +28,18 @@ func (d *ActivityDiagram) LinkAll(s ...shape.Shape) {
 
 // Link places an arrow with a optional label above it between the two
 // shapes.
-func (d *ActivityDiagram) Link(from, to shape.Shape, txt ...string) *shape.Arrow {
+func (d *ActivityDiagram) Link(from, to shape.Shape, action ...string) *shape.Arrow {
 	lnk := shape.NewArrowBetween(from, to)
 	lnk.SetClass("activity-arrow")
 	d.Place(lnk)
-	if len(txt) > 0 {
-		d.placeLabel(lnk, txt[0])
+	if len(action) > 0 {
+		d.placeLabel(lnk, action[0])
 	}
 	return lnk
 }
 
-func (d *ActivityDiagram) placeLabel(lnk *shape.Arrow, txt string) {
-	label := shape.NewLabel(txt)
+func (d *ActivityDiagram) placeLabel(lnk *shape.Arrow, action string) {
+	label := shape.NewLabel(action)
 	switch lnk.Direction() {
 	case shape.DirectionRight, shape.DirectionLeft:
 		d.Place(label).Above(lnk, 20)
@@ -55,46 +57,60 @@ func (d *ActivityDiagram) Place(next ...shape.Shape) *shape.Adjuster {
 	return d.Diagram.Place(next...)
 }
 
-func (d *ActivityDiagram) Then(label string, txt ...string) *shape.Adjuster {
-	next := shape.NewState(label)
+// When adds a new state below the last one
+func (d *ActivityDiagram) Trans(action, label string) *shape.Adjuster {
+	next := shapeFromLabel(label)
 	adj := d.Diagram.Place(next)
 	adj.Below(d.last, d.Spacing)
 	d.VAlignCenter(d.last, next)
-	d.Link(d.last, next, txt...)
+	d.Link(d.last, next, action)
 	d.last = next
 	return adj
 }
 
+// WhenRight adds a new right of last state
+func (d *ActivityDiagram) TransRight(action, label string) *shape.Adjuster {
+	next := shapeFromLabel(label)
+	adj := d.Diagram.Place(next)
+	w := d.Font.TextWidth(action) + d.Spacing
+	adj.RightOf(d.last, w)
+	d.HAlignCenter(d.last, next)
+	d.Link(d.last, next, action)
+	d.last = next
+	return adj
+}
+
+func shapeFromLabel(label string) shape.Shape {
+	if label == "EXIT" {
+		return shape.NewExitDot()
+	}
+	return shape.NewState(label)
+}
+
 // Decide adds a diamond below the last activity
-func (d *ActivityDiagram) Decide(txt ...string) *shape.Diamond {
+func (d *ActivityDiagram) Decide(action ...string) *shape.Diamond {
 	next := shape.NewDecision()
 	adj := d.Diagram.Place(next)
 	adj.Below(d.last, d.Spacing+next.Height()/2)
 	d.VAlignCenter(d.last, next)
-	d.Link(d.last, next, txt...)
+	d.Link(d.last, next, action...)
 	d.last = next
 	return next
 }
 
 // Exit adds an ExitDot below the last activity
-func (d *ActivityDiagram) Exit(txt ...string) *shape.ExitDot {
+func (d *ActivityDiagram) Exit(action ...string) *shape.ExitDot {
 	next := shape.NewExitDot()
 	adj := d.Diagram.Place(next)
 	adj.Below(d.last, d.Spacing)
 	d.VAlignCenter(d.last, next)
-	d.Link(d.last, next, txt...)
+	d.Link(d.last, next, action...)
 	d.last = next
 	return next
 }
 
-// If adds next state to the right with a label.
-func (d *ActivityDiagram) If(after shape.Shape, txt string, next shape.Shape) *shape.Adjuster {
-	adj := d.Diagram.Place(next)
-	adj.RightOf(after, d.Font.TextWidth(txt)+d.Spacing)
-	d.HAlignCenter(after, next)
-	d.Link(after, next, txt)
+func (d *ActivityDiagram) Or(next shape.Shape) {
 	d.last = next
-	return adj
 }
 
 func (d *ActivityDiagram) Start() *shape.Adjuster {