+++ changelog.md
@@ -4,6 +4,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## [unreleased]
+
+- Replace ActivityDiagram.If with Or when switching decision
+- Replace ActivityDiagram.Then with Trans and add TransRight methods
+- Activity diagram default spacing increased from 40 to 60
+
 ## [0.20.0] 2021-04-16
 
 - Add Label.SetHref and Component.SetHref